create fetch profile data on splashscreen
