+++ Profile/ProfileService.swift
@@ -11,6 +11,9 @@ final class ProfileService {
     private let session = URLSession.shared
     private var task: URLSessionTask?
     private var lastToken: String?
+    private(set) var profile: Profile?
+    
+    static let shared = ProfileService()
     
     
     func fetchProfile(_ token: String, completion: @escaping (Result<Profile, Error>) -> Void) {
@@ -24,9 +27,11 @@ final class ProfileService {
             DispatchQueue.main.async {
                 
                 if let error = error {
-                    completion(.failure(error))
-                    self.lastToken = nil
-                    return
+                    DispatchQueue.main.async {
+                        completion(.failure(error))
+                        self.lastToken = nil
+                        return
+                    }
                 }
                 
                 if let response = response as? HTTPURLResponse,
@@ -41,14 +46,22 @@ final class ProfileService {
                 
                 do {
                     let jsonData = try JSONDecoder().decode(ProfileResult.self, from: data)
-                    let profile = self.convertProfileResultToProfile(profileResult: jsonData)
-                    completion(.success(profile))
-                    self.task = nil
-                    
+                    self.profile = self.convertProfileResultToProfile(profileResult: jsonData)
+                    guard let profile = self.profile else {
+                        return
+                    }
+
+                    DispatchQueue.main.async {
+                        completion(.success(profile))
+                        print(profile)
+                        self.lastToken = nil
+
+                    }
                 } catch let error {
-                    completion(.failure(error))
-                    self.lastToken = nil
-                    
+                    DispatchQueue.main.async {
+                        completion(.failure(error))
+                        self.task = nil
+                    }
                 }
             }
         }
@@ -64,7 +77,7 @@ final class ProfileService {
         return request
     }
     
-   private func convertProfileResultToProfile(profileResult: ProfileResult) -> Profile {
+    private func convertProfileResultToProfile(profileResult: ProfileResult) -> Profile {
         return Profile(
             username: profileResult.userName ?? "",
             name: (profileResult.firstName ?? "") + " " + (profileResult.lastName ?? ""),
@@ -73,36 +86,7 @@ final class ProfileService {
         )
     }
     
-    private enum CodingKeysForProfileResult: String, CodingKey {
-        case username, first_name, last_name, bio
-    }
+  
     
-    struct ProfileResult: Codable {
-        let userName: String?
-        let firstName: String?
-        let lastName: String?
-        let bio: String?
-        
-        init(from decoder: Decoder) throws {
-            let container = try decoder.container(keyedBy: CodingKeysForProfileResult.self)
-            userName = try container.decode(String.self, forKey: .username)
-            firstName = try container.decode(String.self, forKey: .first_name)
-            lastName = try container.decode(String.self, forKey: .last_name)
-            bio = try container.decode(String.self, forKey: .bio)
-        }
-    }
-    
-    struct Profile {
-        let username: String
-        let name: String
-        let loginName: String
-        let bio: String
-        
-        init(username: String, name: String, loginName: String, bio: String) {
-            self.username = username
-            self.name = name
-            self.loginName = loginName
-            self.bio = bio
-        }
-    }
+
 }
