+++ SplashViewController/SplashViewController.swift
@@ -11,15 +11,16 @@ import ProgressHUD
 final class SplashViewController: UIViewController {
     
     private let showAuthenticationScreenSegueIdentifier = "ShowAuthenticationScreen"
-    
     private let oAuthService = OAuth2Service()
     private let oAuthTokenStorage = OAuth2TokenStorage()
+    private let profileService = ProfileService.shared
     
     override func viewDidAppear(_ animated: Bool) {
         super.viewDidAppear(animated)
         
         if let token = oAuthTokenStorage.token {
             switchToTabBarController()
+            fetchProfile(token: token)
         } else {
             performSegue(withIdentifier: showAuthenticationScreenSegueIdentifier, sender: nil)
         }
@@ -55,24 +56,47 @@ extension SplashViewController {
 
 extension SplashViewController: AuthViewControllerDelegate {
     func authViewController(_ vc: AuthViewController, didAuthenticateWithCode code: String) {
+        
         UIBlockingProgressHUD.show()
         dismiss(animated: true) { [weak self] in
-            guard let self = self else { return }
-            self.fetchOAuthToken(code)
-            UIBlockingProgressHUD.dismiss()
+            DispatchQueue.main.async {
+                guard let self = self else { return }
+                self.fetchOAuthToken(code)
+                UIBlockingProgressHUD.dismiss()
+            }
         }
     }
     
     private func fetchOAuthToken(_ code: String) {
         oAuthService.fetchOAuthToken(code) { [weak self] result in
-            guard let self = self else { return }
-            switch result {
-            case .success:
-                self.switchToTabBarController()
-            case .failure:
-                break
+            DispatchQueue.main.async {
+                guard let self = self else { return }
+                switch result {
+                case .success(let token):
+                    self.switchToTabBarController()
+                    self.fetchProfile(token: token)
+                case .failure:
+                    // TODO [Sprint 11] Показать ошибку
+                    break
+                }
             }
         }
     }
     
+    private func fetchProfile(token: String) {
+            profileService.fetchProfile(token) { [weak self] result in
+                DispatchQueue.main.async {
+                    guard let self = self else { return }
+                    switch result {
+                    case .success:
+                        UIBlockingProgressHUD.dismiss()
+                        self.switchToTabBarController()
+                    case .failure:
+                        UIBlockingProgressHUD.dismiss()
+                        // TODO [Sprint 11] Показать ошибку
+                        break
+                    }
+                }
+            }
+        }
 }
